Fix camera not following the player
Fixed by putting the "CameraMotor" script on the Main Camera.
